sextractor_md5sum(): turn 'options' parameter into a dictionary
A dictionary is a much more robust data structure than a list here, as
it guarantees that each SExtractor parameter (key) has a corresponding
value. This is less brittle than using a list, where two elements are
needed for each parameter-value pair, and prevents the ugly scenario
in which the user may forget to add to the list either the name of the
parameter or its value, resulting in sometimes-obscure error messages
(for example, if the parameter is missing, SExtractor will assume that
the value is a second FITS image and try to run in double-image mode).