Raise an exception in sextractor() if a configuration file is missing
The latest stable version of SExtractor, 2.8.6, aborts the execution
if either the PARAMETERS_NAME or FILTER_NAME parameters contain the
path to a file that cannot be read. But, in the case of STARNNW_NAME
and the configuration file (-c option), it prints a warning and uses
the internal defaults. In order to treat the four configuration files
equally and make the function more robust, check that they exist and
are readable before calling SExtractor, raising the IOError exception
if any of them is not.