Download sample FITS images from the STScI Digitized Sky Survey
In order to keep the size of the Git repository as small as possible,
the sample FITS images used by some unit tests will be automatically
downloaded from the Digitized Sky Survey. This process is transparent
to the user, which only needs to know what the module-level variable
TEST_IMAGES contains the path to some FITS images.