Add log_uncaught_exceptions() decorator
This decorator catches any uncaught subclass of Exception raised by
the decorated function and logs it with level DEBUG. This one is the
lowest built-in severity level, which means that these messages are
not tracked by default. This decorator is intended to be mainly used,
if not only, to monkey-patch third-party code that raises a non-fatal
exception which we prefer to ignore except when diagnosing problems.