Ignore occasional warning printed by PyRAF's qphot — now for real!
The idea of using a context manager to suppress the warning message
printed to the standard error because of an uncaught exception in
Subprocess.__del__() (commit 251025303f4bc5) turned out to be quite
a blunder: I have learned the hard way that there are no guarantees
regarding when or even whether __del__() will be called. This means
that the warning message is not suppressed if the object happens to
be destroyed outside the with statement.

Use the methods.log_uncaught_exceptions() decorator to monkey-patch
pyraf.subproc.SubprocessError.__del__(), so that (independently of
when it is called), in case an exception is raised it is caught and
the warning message logged with level DEBUG instead of printed to
the standard error.