Add support for multi-extension FITS files to sextractor()
By default, SExtractor runs on all the extensions in a multi-extension
FITS file, which may lead to subtle and hard-to-debug problems. For
example, the FITS images that we download from the STScI Digitized Sky
Survey for the unit tests have a second extension that contains an
(undocumented) table: running SExtractor on it will give us a series
of false detected sources which are, of course, impossible to match
against any reference catalog.

In order to avoid obscure situations like this, and since the images
with which we are going to work are expected to be single-extension,
force SExtractor to always run on a specific extension. Which one is
specified by the new parameter 'ext', which defaults to zero.