Raise TypeError if 'ext' is not an integer
Accepting a floating point number as an index is dangerous, as it is
silently cast to integer and, for example, 3.95 gets truncated to
three. This may be one of those rare cases in which it is not easier
(nor convenient) to ask forgiveness than permission: we better raise
TypeError if the index 'ext' is not an integer, even if it is a float
with integer value, such as 3.0.