Move FITSImage.ahead_file() to the astromatic module
The external header file needed by SCAMP to do astrometry is, not
surprisingly, only used in one function: scamp(). Therefore, there is
not much point in having the code that generates the file as a method
of the FITSImage class — it belongs to the astromatic module, along
with our high-level wrappers of SExtractor, SCAMP and SWarp.