Change unit of parameter 'scale' to arcsec/pixel
The pixel size of the FITS image must be written to the SCAMP external
header file in degrees per pixel, and therefore that is the scale that
ahead_file() expects. However, all the other functions that take the
scale as an argument use arcseconds per pixel, which is more natural
and meaningful to the astronomer. Modify the unit of 'scale' so that
the same one, arcsec/pixel, is used across our pipeline.