Rename check_command() to which()
The current implementation of check_command() does not return the path
of the found executables, but just whether (i.e., True or False) they
are in the current environment. This is not only platform-dependent
(as it calls Unix's 'which' using the subprocess module), but also
makes it impossible to determine the path to the executables: we can
know that they are in the current environment, but not exactly were.

Make the function platform-independent by searching the PATH variable
for the name of the executable and checking whether it exists anywhere
within, returning a list of the full paths to the found executables.
This almost exactly replicates the functionality of Unix's 'which', so
also rename the function to which(). Note that the function can still
be used in the same way as before, as empty lists evaluate to False.

The code has been largely ripped from Twister's repository:
https://twistedmatrix.com/trac/browser/trunk/twisted/python/procutils.py