Mock os.environ to simulate that SExtractor is not installed
The idea of mocking methods.check_command() and make it always return
False in order to test that sextractor() raises SExtractorNotInstalled
when SExtractor cannot be found in the current environment is not the
way to go. It works, but depends on our knowledge that check_command()
is the function that sextractor() uses internally to check whether a
command is installed. This inhibits free refactoring and will cause
the unit test to fail the moment sextractor() no longer uses it.

A rather more appropriate approach to simulate that SExtractor is not
installed is to mock os.environ['PATH'] and set it to an empty string.
In this manner, as the environment variable is empty, all the commands
(and this includes SExtractor) appear as not installed on the system.