Add support for multiple arguments to which()
Allow the function to take multiple arguments, and for each one of
them find the path of the corresponding executables in the current
environment. The returned list contains the paths in the order in
which they were found.

Note that, unlike Unix's 'which' (whose exit status is zero when all
the specified commands are found and executable, and one when one or
more of them is not), this function always succeeds: it returns an
empty list when none of the commands are found, but other than that
there is no way of knowing, without examining the returned list,
whether a specific command was found.