Add the parameter 'options' to scamp()
This parameter is identical to that of sextractor(): a dictionary that
maps each parameter to its value and that overrides their definition
in the SCAMP configuration file (see commit a29f25190e). Note that the
parameters given in this dictionary are only passed to SCAMP: there is
currently no way of overriding the SExtractor parameters when the FITS
LDAC catalog used for astrometry is generated.