_ahead_file(): save external header file to a temporary file
Remove the 'output_path' parameter and instead save the SCAMP external
header file to a temporary file, returning its path. This file is (at
least currently) only used by scamp() and removed as soon as it is no
longer necessary, so so we can save the user from having to determine
the path where it is (temporarily) saved.