Improve the code of scamp()
Restructure the function and rewrite the comments for clarity. Also,
set the MERGEDOUTCAT_TYPE parameter to NONE, so that the calibrated
version of the input SExtractor catalog is not saved by SCAMP. This
'merged' catalog is not used and its only purpose is to be deleted
along with the rest of the contents of the temporary directory, so
we can just avoid creating it in the first place.