Override the CDSCLIENT_EXEC parameter when SCAMP gets executed
The CDSCLIENT_EXEC parameter in the SCAMP configuration file sets the
path to the CDSclient executable that downloads the reference catalog
from a Vizier server. It must be overridden with CDSCLIENT_COMMAND
when SCAMP is run; otherwise, any changes made to this module-level
variable will be ignored — since the value of the parameter will be
that specified in the configuration file.