Do not use the 'with' statement with pyfits.open()
Although we require version 2.0 (released January 2009) or later of
PyFITS, support for the with statement was not added until version 2.3
(May 2010). This means that users with versions older than that get an
AttributeError exception (HDUList' object has no attribute '__exit__')
if the with statement is used when opening a FITS file.