Require SExtractor version 2.8.6 or newer
The syntax that, when working with multi-extension FITS files, makes
it possible to select on which one sources are detected was not added
until SExtractor 2.8.6. Earlier versions interpret the square brackets
surrounding the extension index as part of the name of the FITS file,
which results in error messages such as "cannot open IC5070.fits[0]".
In order words: since commit b1669cd8b0f19 sextractor() does not work
on systems with versions of SExtractor older than 2.8.6.

Add the module-level variable SEXTRACTOR_REQUIRED_VERSION, set it to
the three-element tuple (2, 8, 6) and make sextractor() raise a new
exception, SExtractorUpgradeRequired, if the version installed on the
system is older than this one.