Add Bash completion support for the 'import' command
This is the script that will eventually provide completion support to
our code. It must be copied to /etc/bash_completion.d/ in order to be
automatically sourced, but, for now and until setup.py takes care of
that, when LEMON is installed, it can also be sourced directly (for
example, from your .bashrc). This very first version only completes
the 'import' command and its options.