Do not exit until suggestions (if any) are shown
Move the call to sys.exit() after the piece of code that displays the
suggestions when the command specified by the user is not recognized.
This is, admittedly, pretty obvious (otherwise, these lines of code
will never be run), but we forgot to do it in commit 8f7bba2e6e05d7.