Factor out the creation of the OptionParser instances
The optparse.OptionParser instances used in all of the modules of the
pipeline are always created in the same way: using as description that
of the module and the style.NewlinesFormatter class for printing help
text. Avoid this redundancy with a new module, customparser.py, which
implements the get_parser() function: it takes as its only argument
the description and creates and returns the OptionParser instance.