Do not add a help option to the OptionParser returned by get_parser()
Make optparse not to add the help option (with option strings -h and
--help) to the OptionParser instance that customparser.get_parser()
returns. To print the help text it is enough to type the name of the
command without any arguments, such as `lemon offsets`.

Anyway, even if we wanted to use them, there is the problem that the
--help argument causes 'lemon' to display its own help text and exit,
so there is no way that this option string can be passed down to the
module whose main() function is executed.