Remove short option string of the --follow option from import.py
As per the Zen of Python, explicit is better than implicit, so other
than -v (for --verbose) we are not going to use short option strings
anymore in any of the modules. Since support for Bash completion is
about to be added, users should not even be allowed to complain that
after this they have to type a lot more.