Remove short option strings from seeing.py — and expand long ones
The logic behind this commit is the same that in be5ae81215e52f345ca7:
short option strings are deleted (or, in the case of options with only
a short option string, replaced with a long one), except for -v. Also,
those long option strings whose meaning is not clear enough ('--sper',
for example) are expanded so that they become understandable at first
sight ('--sources-percentile', in this case).