Remove meta-variables from the OptionParsers
The optparse module uses meta-variables to clue in what argument the
user is expected to supply to an option. For example, the meaning of
a description such as "write output to FILE" is much clearer if the
option is listed as "-f FILE".

However, in the case of our modules an overwhelming majority of the
meta-variables are redundant: things like "-percentile=PERCENTILE" do
not add any useful information and only seem to serve to decrease the
space that is available to display the help text of the options.

There does not exist a way to disable meta-variables, but if they are
explicitly set to a string which only contains a whitespace the help
text will display "-percentile= ", getting rid of them for all
practical purposes. Use the new function customparser.clear_metavars()
to set all the meta-variables to a whitespace — this saves us from
having to modify the tens of already-existing calls to add_option().