Add missing else clause in FITSImage.date()
No matter what the result of the previous statements is, the missing
else means that we always attempt to match the date of observation of
the image to the last regular expression. In other words: the success
of the method is exclusively determined by whether the last regex
('yyyy-mm-ddTHH:MM:SS[.sss]') matches, while the first two (formats
'yy/mm/dd' and 'yyyy-mm-dd') are ignored.