Avoid empty subdirectories of automatically discarded FITS images
Postpone the creation of the subdirectories where images that are
automatically discarded are saved until we can confirm that at least
one image is effectively going to be so. In this manner, we avoid
ending up with empty directories if for any of the criteria (FWHM
and elongation) nothing is discarded.