Set up the Sphinx source directory
After several months postponing this task, the time has finally come
to document LEMON — we are slowly getting dangerously close to the
commissioning of PANIC and a proper user guide is a must. We will use
Sphinx, of course, an elegant weapon venerated by Python developers.

This commit is just the result of running sphinx-quickstart to create
the documentation source directory, Doc, with the configuration file
conf.py and a master document, index.html, which serves as a welcome
page and to contain the root of the Table of Contents tree.