Add --tempdir option
This option allows the user to make a copy of the temporary aligned
images that were combined into the mosaic. These files have been so
far removed when the program exits, deemed to be necessary only in
order to create the mosaic. However, they can be very useful not just
for debugging, but also if astronomers want to make sure that all the
images were indeed aligned correctly.

Many thanks to Javier Blasco for suggesting this, and also for sharing
with us such a useful tip: DS9, the astronomical imaging and data
visualization application, can open multiple FITS files from the
command line, each one in a different frame: `ds9 *.fits`. The 'blink'
option can then be used to cycle through all the available frames,
allowing us to visually confirm that the alignments are correct.