Add FITSImage.pfilter()
This is a convenience method that reads a keyword from the header of
the FITS image and returns it as a Passband object. This allows us to
get the photometric filter of an image directly as a Passband object,
which in most cases is what we need. Additionally, this approach has
the advantage that, if the name of the filter is unknown to the
Passband class, the error message of the raised exception,
NonRecognizedPassband, includes the filename of the FITS file.