Use the name of the original images as prefix for the temporary files
The filename of the temporary aligned images was irrelevant when they
were always removed after the mosaic was created, but now that the
--tempdir option is available (commit 7769bfea4efe161ae4b4c8) we want
them to have names more meaningful than, e.g., "tmpF9rV7J.fits". Make
them be related to the filename of the original image by using it as
the prefix of the temporary file, so that it becomes something like
"cig0181_H6652_aligned_F9rV7J.fits".