Store the pixel shift of the aligned images in HISTORY keywords
For debugging purposes, store in the header of the temporary aligned
FITS images the pixel shift in both the x- and y-axes, using HISTORY
keywords. The process of aligning images with LEMON is not as simple
as applying these offsets, nevertheless, as in order not to lose any
data blank margins are added to the images before the alignment. For
this reason, the width of these margins is also stored in additional
HISTORY's.