Retain the original filename when --filename is an empty string
The --filename option sets the filename with which the 'best-seeing'
image is copied to the output directory. However, it may be the case
that the user is interested in simply running SExtractor on all of
the files without treating any of them differently. Implement this
functionality by changing --filename so that, when set to an empty
string, the image with the best FWHM retains its original filename.

Kudos to Javier Blasco for this suggestion.