Use datetime.strptime() instead of time.strptime()
I have just noticed that the time.gmtime() function, which converts a
time expressed in seconds since the epoch to a struct_time in UTC,
ignores fractions of a second. Although the Python documentation does
not mention why, the reason is that struct_time does not store these
fractions: tm_sec is defined as an integer in the range [0, 61].

As pointed out by a Stack Overflow user [1], we can use a datetime
object, which since version 2.6 supports a %f format code that expands
to the number of microseconds. We still have to use a struct_time, as
input to datetime.utctimetuple(), but the method datetime.strftime()
now allows us to easily extract the fraction and add it to the time
expressed in seconds.

[1] http://stackoverflow.com/a/7000121/184363