FITSImage.date(): accept all the formats defined in the FITS Standard
The FITS Standard defines two date formats: the old one, 'yy/mm/dd',
(deprecated, may be used only for dates from 1900 through 1999) and
the new Y2K-compliant one: 'yyyy-mm-dd[THH:MM:SS[.sss]]'. However,
our code has until now supported exclusively the latter format, and
only in its form 'yyyy-mm-ddTHH:MM:SS[.sss]'.

Improve FITSImage.date() so that it accepts and parses correctly all
the formats defined in the standard. The most noticeable change is
that the method now receives a third keyword argument, 'time_keyword',
the name of the FITS keyword (almost always, TIME-OBS) from which the
time at the start of the observation is read when it is not included
directly as part of the value of the 'date_keyword' keyword (that is,
DATE-OBS). In other words: if the format of the date is 'yy/mm/dd' or
'yyyy-mm-dd', the time must be in the 'time_keyword' keyword, with
the format 'HH:MM:SS[.sss]'.