Add the --timek option
The three LEMON commands that use the date of the observation of the
FITS images ('import', 'astrometry' and 'offsets') need a new option,
--timek (default: TIME-OBS), to specify the FITS keyword that contains
the time at the start of the observation (see commit 30f99e69d72f816c).
This keyword is needed if the time is not included directly as part of
the --datek keyword value with the format 'yyyy-mm-ddThh:mm:ss[.sss]'.

Note that the 'astrometry' command does not call the FITSImage.date()
method, but it needs --datek and --timek in order to propagate them,
among several other FITS keywords, from the input FITS image to the
astrometrically-calibrated output one.