Swap x and y values of the size in numpy.random.random_integers()
The shape of the matrix of random integers that is created with the
numpy.random.random_integers() function must be (y, x), not (x, y).
In other words, the number of columns of the FITS image must be given
first, because PyFITS indexes pixels by column rather than by row.