Fix two exceptions incorrectly raised by FITSImage.__init__()
There are two scenarios in which FITSImage.__init__() does not raise
the appropriate exception. On the one hand, when the 'SIMPLE' keyword
is missing from the primary header of the FITS file, NonStandardFITS,
instead of NonFITSFile, should be raised. On the other hand, we raise
this exception, NonFITSFile, when the file is not readable, but the
proper one is IOError.