Add getter for LEMONdB.mosaic attribute
We are going to use a new table, RAW_IMAGES, to store as a blob entire
FITS files. The ID must refer to that of the file in the IMAGES table.
As the reference frame is stored in its own table, RIMAGE, the special
value zero is used as the ID for its FITS file.