Add 'Finding Chart' button to the toolbar
The button does not yet do anything, although it already becomes
enabled when a LEMONdB is opened and is disabled if it is closed.

The compass icon is part of the Isloo iconset released by IconFinder
(http://www.iconfinder.com/iconsets/Isloo_icons_by_IconFinder) in May
2010 and was designed by Asher Abbasi (http://kyo-tux.deviantart.com).
License: Creative Commons Attribution-No Derivative Works 3.0.