Store the reference FITS image in the LEMONdB
Use the LEMONdB.mosaic attribute setter to store the reference FITS
image (that on which sources are detected, and from which offsets are
calculated) in the output LEMONdB. In an ideal world we would store in
the database all the reduced images, but in practice this would easily
scale the size of the LEMONdB up to the hundreds of gigabytes in the
case of long campaigns. The reference image, being the only one we
need to display a finding chart in the GUI, is enough for now.