Use APLpy to display the finding chart
Connect the 'Finding Chart' button to the view_finding_chart()
handler, which displays in a GTK.Dialog the reference image using
APLpy (the Astronomical Plotting Library in Python). An object of the
FindingChartDialog class is created every time the button is clicked,
which is suboptimal — but sufficient for a first version.