Hide (not destroy) the Finding Chart window when it is closed
If, instead of destroy()'ing it, we hide() the dialog when the run()
method exits, we can run() it again as many times as we need to. This
allows us to create a FindingChartDialog object only once, the first
time the button is clicked, and hide and show it repeatedly as the
user opens and closes the Finding Chart window. In this manner the
FITS file must be read by APLpy only once and the window remembers
its size, position and any Matplotlib panning and zooming.

Thanks to user 'ptomato' on Stack Overflow for this explanation:
http://stackoverflow.com/a/4663232/184363