Add FindingChartDialog.mark_star()
When the user clicks on the finding chart, find the star closest to
those x- and y-image coordinates and highlight it using a red marker.
This functionality is implemented by FindingChartDialog.mark_star(), a
callback function to which Matplotlib's 'button_press_event' (mouse
button is pressed) is connected. The markers are always stored in a
new layer with the same name, so that only the last one is visible.