Add a 'Go to Star' button to the Finding Chart window
Use a stock button, STOCK_GO_FORWARD (an arrow pointing to the right)
with a different label, 'Go to Star'. In order to achieve this, as
suggested by Christian Reis in the PyGTK FAQ, we get the label child
of the stock button and set the text directly. The button is not yet
functional, but it is already disabled until a star is selected.