Split LEMONJuicerGUI.handle_row_activated() into two methods
Factor out part of LEMONJuicerGUI.handle_row_activated() into a new
method, LEMONJuicerGUI.view_star() — which takes care of adding a new
page to the gtk.Notebook with the details of the star, or switches to
it if the page already exists. This allows us to view the details of
a star by calling this method anywhere in our code, and not only by
clicking in the list of stars.