Make the Finding Chart window non-modal
By making the Finding Chart non-modal and, therefore, not preventing
the workflow on the main window, the user can arrange the two windows
side by side, easily switching between them. The run() method blocks
in a recursive main loop until the dialog either emits the 'response'
signal or is destroyed, so we now need to use show().