Make the <Control>F accelerator work in the Finding Chart window too
Setting <Control>F as the accelerator of the 'Finding Chart' button,
in the toolbar of the LEMONJuicerGUI, has the limitation that it does
not work when the focus is on a different window. However, in the same
way that clicking the button again closes the Finding Chart window, we
would like to be able to show and hide it by repeatedly pressing the
accelerator key.

To achieve this, install the <Control>F accelerator in an accelerator
group in the main window of LEMONJuicerGUI, and then connect to it the
FindingChartDialog window. Since both windows are associated with the
same accelerator group, <Control>F works independently of where the
focus is.