Turn the 'Finding Chart' button into a gtk.ToggleToolButton
Use a gtk.ToggleToolButton, not a gtk.ToolButton, for the 'Finding
Chart' button in the LEMONJuicerGUI toolbar, and update its status to
reflect whether the Finding Chart window is being displayed: active if
visible, inactive if not.

However, this window can be shown and hidden from other events than
just clicking the button (namely, the 'Finding Chart' item in the View
menu and the <Control>F accelerator), so when we update the status of
the button we need to temporarily suppress the 'toggled' signal that
is emitted when gtk.ToggleToolButton.set_active() is called. That is
what FindingChartDialog.set_finding_chart_button_active() does.