Set 'Go to Star' as the default button when a star is selected
When a star is selected in the Finding Chart window, change the
default button (normally activated pressing Enter) from 'Close' to 'Go
to Star'. If the user clicks the button, then set 'Close' again as the
default response. This allows to rapidly view the details of a star
and then close the Finding Chart window by pressing Enter twice.

Regardless of the above paragraph, the default button is always set to
'Close' when the Finding Chart window changes from hidden to visible,
even although a star could have been previously selected. This marker
is not deleted because it acts as a useful reminder of what was the
last selected star.