Mark a star when the user clicks on 'View in Finding Chart'
Add a new method, StarDetailsGUI.handle_view_star_in_chart(), and
connect to it the 'clicked' signal of the 'View in Finding Chart'
button. This method shows the gtk.Dialog, if not already visible, and
in turn calls the FindingChartDialog.mark_star() method, overlaying a
green marker (#24ff29) on the star.