Fix the 'Save' button on the matplotlib navigation toolbars
Although the matplotlib navigation toolbars under the light curves and
the finding chart work, allowing the user to navigate through the data
set, clicking on the 'Save' button raises the "TypeError: parent must
be a GtkWindow or None" exception. The reason for this is that we are
passing a gtk.HBox object as the second argument to the __init__()
method of the NavigationToolbar class. Use gtk.Widget.get_window() to
get the widget's gtk.gdk.Window that NavigationToolbar expects.