Suggest relevant filenames in the 'Save' button gtk.FileChooserDialog
The filename that is suggested by the gtk.FileChooserDialog when the
user clicks on the 'Save' button on the matplotlib navigation toolbar
is the string returned by FigureCanvas.get_window_title() or, in case
there is no window containing the figure, a generic 'image'.

Monkey-patch FigureCanvas.get_window_title() so that, instead of the
vague 'image', the file chooser suggests a relevant filename: (a) the
name of the observed field (as returned by LEMONdB.field_name), if we
are saving the finding chart, and (b) a string containing the name of
the field, the ID of the star and the photometric filter, in the case
of light curve plots.