Add the --individual option
Make it possible to do photometry using aperture and sky annuli
derived from the FWHM of each image, instead of using the median FWHM
of all the images in the same filter. As concisely explained by Javier
Blasco in issue #4, a fixed aperture may enclose different percentages
of the stars, and therefore lead to suboptimal signal-to-noise ratios,
when the seeing is variable and the aperture small.