Add Travis YAML file
Travis CI is a continuous integration service, which automatically
detects when a commit has been pushed to the GitHub repository and,
when this happens, attempts to build the project and run the tests.

The installation of the LEMON dependencies is not a trivial task, but,
although crudely, .travis/travis-setup.sh attempts to automate the
process. Travis runs builds in isolated virtual machines that offer a
vanilla build environment (i.e., no state persists between builds),
so, until IRAF can be straightforwardly installed using dpkg or pip,
we do not have to worry that the script is not an elegant solution.