Install Python packages using pip
Using pip to install all the Python dependencies allows us to specify
the exact version, or the minimum one, required for each package. It
also makes that part of the installation instructions applicable to
non-Debian based operating systems, where APT is not available. In
fact, PANIC itself is expected to run on openSUSE 12.2, so we will
also benefit from this change in a few months, when the instrument
is commissioned at the Calar Alto Observatory.