Remove NonFITSFile exception — always use NonStandardFITS
The line between "FITS files that do not follow the standard" and
"files that are not in the FITS format" is too vague sometimes (for
example, if we come across a heavily corrupted FITS file), and the
benefit of improving our code so that the right exception is always
raised is almost certainly not worth the effort. Use NonStandardFITS
always: after all, an empty, corrupted or non-FITS file can also be
considered not to follow the standard.