Use apt-get again to install several Python dependencies
Forget what we said in commit 6e2283a0523daca29d1d885da6e. Downloading
the binary packages with apt-get is tremendously faster than compiling
lots of C code with pip every time the Travis tests are run. We can do
this because, unlike in Debian GNU/Linux, the versions of the packages
available in Ubuntu 12.04 are recent enough to be used to run our unit
tests.