Test against Python 2.7 only
Although 'system_site_packages' is set to 'true' in travis.yml, the
Python 2.6 virtual environment cannot see the packages installed with
apt-get. The reason for this is that on Ubuntu 12.04 system packages
only include Python 2.7 libraries. Thanks so much to Josh Kalderimis
and Mathias Meyer, from the Travis CI support team, for spinning up
a debug virtual machine and helping me figure this out.