Use TextTestRunner instead of unittest.main()
The 'exit' parameter of unittest.main(), which, when set to False,
displays the result on standard output without calling sys.exit(), was
not added until Python 2.7, but we need our code to be 2.6 compatible.
TestLoader and TextTestRunner do not suffer from this problem, and are
in fact a preferable way to run the tests, since they give us a much
finer level of control.