Use os.path.basename() on sys.argv[0]
It is operating system dependent whether sys.argv[0], the script name,
is a full pathname or not. Use os.path.basename() to make sure that
only the name of the script, 'lemon', is always printed by the help
message, regardless of where it is executed from — in other words,
avoid coming across messages like "See '/home/vterron/lemon/lemon
COMMAND' for more information on a specific command.").