Use os.path.abspath() on __file__
__file__ returns the filename used to locate and load a module. When
the module is found in the current directory, we just get a relative
path, such as 'fitsimage.py'. This makes os.path.dirname() return an
empty string, which then causes methods.tmp_chdir() to raise "OSError:
[Errno 2] No such file or directory:" when it attempts to os.chdir()
to ''. Turn the path into an absolute one so that our code does not
break when fitsimage.py or qphot.py are imported into the interpreter.