Turn all the values in Passband.SYSTEM_LETTERS into tuples
When parsing the name of a filter, we will need to be able to check
whether the letter belongs to that system. For example, "Gunn R" is
valid because 'R' is one of 'UVGR'. The problem with using a string,
however, is that substrings are also considered valid letters (e.g.,
'UV' in 'UVGR' returns True). Avoid this by converting these strings
to tuples, such as ('U', 'V', 'G', 'R').