Fix bug in LEMONdBTest.test_pfilters_and_star_pfilters()
There are two errors in test_pfilters_and_star_pfilters(). First, we
are not testing that the returned list of Passband objects is sorted,
something that is always the case according to the docstring of the
Passband.pfilters() method. Second, and even worse, we are mistakenly
using TestCase.assertTrue() instead of assertEqual().