Remove outdated references to the NonFITSFile exception
The NonFITSFile exception was removed in commit 8ac66a4e5be351b5b6,
but there are still a few references to it in other modules, which
eventually causes the AttributeError ("'module' object has no
attribute 'NonFITSFile'") exception to be raised. Thanks to Pablo
Ramírez for letting us know about this problem.