Fix bug in which() — consider only executable files
The function which() does not test whether the executables in PATH
correspond to files. Therefore, directories with execute permissions
may also be incorrectly identified as commands — e.g., which('test')
will return the path to the unit tests directory, test/, in case it
is executable, since as per our installation instructions the LEMON
directory is added to the PATH. Use os.path.isfile() to avoid this.