Use a 200-pixel width lemon image in the README
As explained by the staff, GitHub strips any inline CSS, so the
explicit width of the lemon image, specified in the README file with
the reStructuredText :width: option, does not make it through to get
rendered. Until image resizing is supported, the only workaround is to
use an image already sized to the dimensions we want. Thanks to Robert
Sese, from the GitHub support team, for the helpful explanation on why
the image was not displaying correctly.