Set 'iraf' and 'IRAFARCH' environment variables in Travis build
The Travis CI build seems somehow to be unable to get the 'iraf' and
'IRAFARCH' environment variables from the CL script, which causes the
build to fail with the following error when PyRAF is imported:

  Your iraf and IRAFARCH environment variables are not defined and
  could not be determined from /usr/local/bin/cl. Before starting
  pyraf, define them by doing (for example)

  setenv iraf /usr/local/iraf/
  setenv IRAFARCH redhat

  at the Unix command line. The values will depend on your IRAF
  installation.

There are also a couple of csh "Word too long" error messages. PyRAF
is perfectly capable of determining the value of both environment
variables in all the machines on which we have installed it so far, so
I assume there must be something wrong with the IRAF configuration in
the Travis VM. Without SSH access I am unable to figure out the cause
of this error: fix it by manually setting both environment variables
before running our test suite.