Make sure the output FITS files are user-writable
shutil.copy2() copies the file as well as the metadata, so we need to
make sure that the output images are user-writable — otherwise, the
"IOError: [Errno 13] Permission denied" exception is raised when we
attempt to update their FITS headers. Although this `chmod u+w`'ing is
already done by the 'import' command, our users may not always use it,
starting the data reduction directly with 'seeing'.