Add a trigger to make sure there is only one sources image
LEMONdB.add_image() overrides the previous (if any) sources image when
'_is_sources_img' is True, but we'd rather err on the side of caution.
Add a trigger so that sqlite3.IntegrityError is raised if at any point
there is more than one row in the IMAGES table with SOURCES == 1.