Allow NULL column values only for the sources image
Add a trigger to raise sqlite3.IntegrityError if any of the FILTER_ID,
UNIX_TIME, AIRMASS or GAIN columns are NULL for 'normal' images (those
on which we do photometry, i.e., SOURCES == 0). These columns are only
optional for the sources image. This is because astronomical sources
may be detected on a mosaic generated with the IPAC's Montage toolkit
and, in those cases, we may be combining multiple photometric filters,
dates of observation, airmasses and gains. Note that the OBJECT column
may always be NULL, regardless of the value of SOURCES.