Add LEMONdB.vmin property
This property stores the 'vmin' parameter of APLpy's logarithmic
normalization algorithm (stretch = 'log') in the METADATA table. This
value can be reused by the LEMON Juicer GUI in order to normalize the
finding chart — so that the appropriate value of the parameter has to
be identified only the first time we work with the image.