Fix bug in LEMONdB.add_image()
The DuplicateImageError exception is not raised when there already
is an image in the database with the same Unix time and photometric
filter, but instead if there is (a) an image with the same Unix time
and (b) an image (not necessarily the same one) with the same filter.
This bug has gone unnoticed because so far the try-except block only
raises an exception when we come across a duplicate image.