Use PyFITS 3.2 — stay away from what looks like a bug in v3.3
The Travis CI build is failing after the release of PyFITS 3.3, since
the FITSImage class is apparently reading an incorrect number of HDU
objects. This seems to happen only in 3.3; in older versions the code
has worked as expected. I have opened an issue letting the developers
know — until we understand what is wrong and, in case it is indeed a
bug in PyFITS, it gets fixed, we should prudently stay away from 3.3.

Incorrect number of HDU objects if 'SIMPLE' is missing:
https://github.com/spacetelescope/PyFITS/issues/94